refactor: split firestore logic into domain services
