+++ src/app/core/services/firestore.service.ts
@@ -19,14 +19,10 @@
 ***********************************************************************************************/
 // src/app/core/services/firestore.service.ts
 import {Injectable} from "@angular/core";
-import {combineLatest, firstValueFrom, Observable, of} from "rxjs";
+import {Observable, of} from "rxjs";
 import {map, catchError} from "rxjs/operators";
 import {FirebaseError} from "firebase/app";
-import {Account, RelatedAccount} from "@shared/models/account.model";
-import {AngularFirestore, Query} from "@angular/fire/compat/firestore";
-import {AngularFireStorage} from "@angular/fire/compat/storage";
-import {RelatedListing} from "@shared/models/related-listing.model";
-import {DocumentData, WhereFilterOp} from "firebase/firestore";
+import {AngularFirestore} from "@angular/fire/compat/firestore";
 
 @Injectable({
   providedIn: "root",
@@ -34,7 +30,6 @@ import {DocumentData, WhereFilterOp} from "firebase/firestore";
 export class FirestoreService {
   constructor(
     private afs: AngularFirestore,
-    private storage: AngularFireStorage,
   ) {}
 
   // Helper to ensure documents have an id property
@@ -190,230 +185,4 @@ export class FirestoreService {
     }
   }
 
-  /**
-   * Retrieves documents from a collection based on a specified condition in real-time.
-   *
-   * @param {string} collectionName - Name of the collection.
-   * @param {string} field - Field name to apply the condition on.
-   * @param {any} condition - Condition to apply.
-   * @param {any} value - Value to match against.
-   * @returns {Observable<T[]>} - Returns an Observable of an array of documents that match the condition.
-   */
-  getCollectionWithCondition<T>(
-    collectionName: string,
-    field: string,
-    condition: any,
-    value: any,
-  ): Observable<T[]> {
-    const collectionRef = this.afs.collection<T>(collectionName, (ref) =>
-      ref.where(field, condition, value),
-    );
-    return collectionRef.snapshotChanges().pipe(
-      map((actions) =>
-        actions.map((action) => {
-          const data = action.payload.doc.data() as T;
-          const id = action.payload.doc.id;
-          return this.populateId(data, id); // Ensure id is populated
-        }),
-      ),
-      catchError((error) => {
-        console.error("Error retrieving collection with condition:", error);
-        return of([]);
-      }),
-    );
-  }
-
-  /**
-   * Searches for documents in a collection based on a name field in real-time.
-   *
-   * @param {string} collectionName - Name of the collection.
-   * @param {string} searchTerm - Term to search for.
-   * @param {string} [userId] - Current user Id (optional).
-   * @returns {Observable<Account[]>} - Returns an Observable of an array of accounts that match the search term.
-   */
-  searchAccountByName(
-    collectionName: string,
-    searchTerm: string,
-  ): Observable<Account[]> {
-    const collectionRef = this.afs.collection<Account>(
-      collectionName,
-      (ref) => {
-        let queryRef = ref
-          .where("privacy", "==", "public")
-          .where("type", "in", ["user", "group"]);
-
-        // Firestore does not support OR queries directly, but we can perform multiple queries and combine the results
-        // For simplicity, we'll only handle the search term as-is
-        queryRef = queryRef
-          .orderBy("name")
-          .startAt(searchTerm)
-          .endAt(searchTerm + "\uf8ff");
-
-        return queryRef;
-      },
-    );
-
-    return collectionRef.valueChanges({idField: "id"}).pipe(
-      catchError((error) => {
-        console.error("Error searching accounts:", error);
-        return of([]);
-      }),
-    );
-  }
-
-  /**
-   * Retrieves documents from a relatedAccounts sub-collection for a given account in real-time.
-   *
-   * @param {string} accountId - ID of the account.
-   * @returns {Observable<RelatedAccount[]>} - Returns an Observable of an array of related accounts.
-   */
-  getRelatedAccounts(accountId: string): Observable<RelatedAccount[]> {
-    const relatedAccountsRef = this.afs.collection<RelatedAccount>(
-      `accounts/${accountId}/relatedAccounts`,
-    );
-
-    return relatedAccountsRef.snapshotChanges().pipe(
-      map((actions) =>
-        actions.map((action) => {
-          const data = action.payload.doc.data() as RelatedAccount;
-          const id = action.payload.doc.id;
-          return this.populateId(data, id); // Ensure id is populated
-        }),
-      ),
-      catchError((error) => {
-        console.error("Error getting related accounts:", error);
-        return of([]);
-      }),
-    );
-  }
-
-  /**
-   * Retrieves all public user or group accounts from the 'accounts' collection in real-time.
-   * @returns {Observable<Account[]>} - Returns an Observable of an array of account documents.
-   */
-  getAllAccounts(): Observable<Account[]> {
-    return this.afs
-      .collection<Account>("accounts", (ref) =>
-        ref
-          .where("privacy", "==", "public")
-          .where("type", "in", ["user", "group"]),
-      )
-      .snapshotChanges()
-      .pipe(
-        map((actions) =>
-          actions.map((action) => {
-            const data = action.payload.doc.data() as Account;
-            const id = action.payload.doc.id;
-            return this.populateId(data, id); // Ensure id is populated
-          }),
-        ),
-        catchError((error) => {
-          console.error("Error getting accounts:", error);
-          return of([]);
-        }),
-      );
-  }
-
-  /**
-   * Retrieves an account with its related accounts and related listings.
-   *
-   * @param {string} accountId - ID of the account.
-   * @returns {Observable<{ account: Account | null; relatedAccounts: RelatedAccount[]; relatedListings: RelatedListing[] }>}
-   */
-  getAccountWithRelated(accountId: string): Observable<{
-    account: Account | null;
-    relatedAccounts: RelatedAccount[];
-    relatedListings: RelatedListing[];
-  }> {
-    return combineLatest([
-      this.getDocument<Account>("accounts", accountId),
-      this.getRelatedAccounts(accountId),
-      this.afs
-        .collection<RelatedListing>(`accounts/${accountId}/relatedListings`)
-        .snapshotChanges()
-        .pipe(
-          map((actions) =>
-            actions.map((action) => {
-              const data = action.payload.doc.data() as RelatedListing;
-              const id = action.payload.doc.id;
-              return this.populateId(data, id); // Ensure id is populated
-            }),
-          ),
-          catchError(() => of([])),
-        ),
-    ]).pipe(
-      map(([account, relatedAccounts, relatedListings]) => ({
-        account,
-        relatedAccounts,
-        relatedListings,
-      })),
-      catchError((error) => {
-        console.error("Error getting account with related data:", error);
-        return of({
-          account: null,
-          relatedAccounts: [],
-          relatedListings: [],
-        });
-      }),
-    );
-  }
-
-  // Firebase Query Logic (Ends) //
-
-  /**
-   * Retrieves documents from any collection or sub-collection.
-   * @param fullPath The full Firestore path to the collection (e.g., 'listings/{listingId}/relatedAccounts').
-   * @param conditions Optional conditions to filter documents.
-   * @returns {Observable<T[]>}
-   */
-  getDocuments<T>(
-    fullPath: string,
-    conditions?: {
-      field: string;
-      operator: WhereFilterOp;
-      value: any;
-    }[],
-  ): Observable<T[]> {
-    const collectionRef = this.afs.collection<T>(fullPath, (ref) => {
-      let query: Query<DocumentData> = ref; // Use Query type
-      if (conditions) {
-        conditions.forEach((condition) => {
-          query = query.where(
-            condition.field,
-            condition.operator,
-            condition.value,
-          );
-        });
-      }
-      return query;
-    });
-
-    return collectionRef.snapshotChanges().pipe(
-      map((actions) =>
-        actions.map((action) => {
-          const data = action.payload.doc.data() as T;
-          const id = action.payload.doc.id;
-          return {...data, id}; // Ensure id is populated
-        }),
-      ),
-      catchError((error) => {
-        console.error(
-          `Error retrieving documents from path: ${fullPath}`,
-          error,
-        );
-        return of([]);
-      }),
-    );
-  }
-
-  // Upload a file to Firestore storage
-  async uploadFile(path: string, file: File): Promise<string> {
-    const fileRef = this.storage.ref(path);
-
-    // Upload the file
-    await this.storage.upload(path, file);
-
-    // Get the download URL using firstValueFrom
-    return firstValueFrom(fileRef.getDownloadURL());
-  }
 }
